Crate bevy_mod_raycast

Modules

rays

Encapsulates Ray3D, preventing use of struct literal syntax. This allows us to guarantee that the Ray3d direction is normalized, because it can only be instantiated with the constructor.

Structs

DebugCursor

DebugCursorMesh

DebugCursorTail

DefaultPluginState

Global plugin state used to enable or disable all ray casting for a given type T.

DefaultRaycastingPlugin

Intersection

Holds computed intersection information

Ray3d

A 3D ray, with an origin and direction. The direction is guaranteed to be normalized.

RayCastMesh

Marks an entity as pickable, with type T.

RayCastSource

The RayCastSource component is used to generate rays with the specified cast_method. A ray is generated when the RayCastSource is initialized, either by waiting for update_raycast system to process the ray, or by using a with_ray function.

SimplifiedMesh

Triangle

Enums

Primitive3d

RayCastMethod

Specifies the method used to generate rays.

RaycastSystem

Traits

IntoUsize

TriangleTrait

Functions

build_rays

ray_intersection_over_mesh

Cast a ray on a mesh, and returns the intersection

ray_mesh_intersection

Checks if a ray intersects a mesh, and returns the nearest intersection if one exists.

update_debug_cursor

Updates the 3d cursor to be in the pointed world coordinates

update_raycast

Iterates through all entities with the RayCastMesh component, checking for intersections. If these entities have bounding volumes, these will be checked first, greatly accelerating the process.